fix blt

+++ jolt-core/src/jolt/instruction/blt.rs
@@ -62,7 +62,7 @@ impl JoltInstruction for BLTInstruction {
     }
 
     fn lookup_entry_u64(&self) -> u64 {
-        (self.0 < self.1).into()
+        ((self.0 as i32) < (self.1 as i32)).into()
     }
 
     fn random(&self, rng: &mut StdRng) -> Self {
@@ -105,7 +105,6 @@ mod test {
             );
         }
     }
-
 
     use crate::jolt::instruction::test::{lookup_entry_u64_parity_random, lookup_entry_u64_parity};
